Added MobileBarcodeScanner Constructor with UIViewController parameter
I added a MobileBarcodeScanner(object delegateController) constructor
to enable developers to pass along a UIViewController that will
function as the delegate for the scanner. This is handy when you, for
instance, are using a Storyboard and cannot set the
window.RootViewController because this is loaded from the Storyboard in
execution time.