Android Multi Camera Support
For applications targeting API Level 9 or higher, the library will now
attempt to find a Rear camera to use.  If there is no rear camera, it
will use the first available camera, in the case of devices with only a
front facing camera (like nexus 7), this will be the camera selected.
Fixes #3