Release Holder callback on Android
When the surface is destroyed, we should release the holder’s callback (as per #617).

However we also need to add the holder back in when the view is attached to the window again.  I’ve added some code for this scenario, and some checks to make sure we aren’t adding multiple callbacks.

This fixes #617. Thanks!