iOS: Keep track of and rotate buffer if necessary
This is basing off the status bar orientation, which I think should be good enough.  This basically ensures that if we are not using AutoRotate = true, we still rotate the preview buffer to the right orientation so scanning can work on 1D barcodes.

Really upside down or right side up is enough, so we just keep track of if we’re in portrait or portrait upside down and then rotate once if so (for whatever reason the preview buffer image data comes out rotated by default).