Return result after Stop
We are currently using ZXing.Net.Mobile to scan barcodes for an internal project. When we integrated the package, we found a bug in conjunction with another open source package (Acr.UserDialogs). When we showed a Dialog after scanning a QR Code, the overlay was shown on the QR Code UI, which was immediatly closed afterwards. For that reason I changed the ZXing.Net.Mobile package to pass the result of the scan back, after the UI was closed.